server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.2
google-genai==1.5.0
python-dotenv==1.0.1
aiofiles==24.1.0
gitpython==3.1.44
